dbeaver/pro#6042 get node features from feature provider
