+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/registry/WebObjectFeatureProviderDescriptor.java
@@ -1,0 +1,39 @@
+/*
+ * DBeaver - Universal Database Manager
+ * Copyright (C) 2010-2024 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.cloudbeaver.registry;
+
+import io.cloudbeaver.service.navigator.DBWFeatureProvider;
+import org.eclipse.core.runtime.IConfigurationElement;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.model.impl.AbstractDescriptor;
+
+public class WebObjectFeatureProviderDescriptor extends AbstractDescriptor {
+
+    private final String id;
+    private final DBWFeatureProvider instance;
+
+    public WebObjectFeatureProviderDescriptor(IConfigurationElement cfg) throws DBException {
+        super(cfg);
+        this.id = cfg.getAttribute("id");
+        ObjectType implClass = new ObjectType(cfg.getAttribute("class"));
+        this.instance = implClass.createInstance(DBWFeatureProvider.class);
+    }
+
+    public DBWFeatureProvider getInstance() {
+        return instance;
+    }
+}
